requirements.txt
# Deadlock Bot – Python dependencies
# Generated from .venv (2026-02-18) — only direct project dependencies

# Discord
discord.py==2.7.1

# HTTP / Dashboard
aiohttp==3.13.5
aiohttp-jinja2==1.6
aiohttp-session==2.12.1
aiohttp-basicauth==1.1.0

# Database
aiosqlite==0.22.1
psycopg[binary]==3.3.3

# Templating
Jinja2==3.1.6

# Data / validation
pydantic==2.12.5
pydantic-settings==2.13.1
pytz==2026.1.post1

# HTTP client
httpx==0.28.1
requests==2.32.5

# HTML parsing
beautifulsoup4==4.14.3

# Crypto / auth
cryptography==46.0.5
keyring==25.7.0
